Implementing Main API gateways, Faking data

+++ Backend/Java/booktalk-app/src/main/java/com/booktalk/booktalk_app/Entities/User.java
@@ -47,7 +47,7 @@ public class User {
 
     @Column(nullable = false)
     @NotNull
-    private Integer frequency;
+    private Double frequency;
 
     @OneToMany(mappedBy = "user")
     private List<UserBook> connectors;
